requirements.txt
black==21.12b0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
idna==3.3
munch==2.5.0
mypy-extensions==0.4.3
numpy==1.21.5
pandas==1.3.5
pathspec==0.9.0
patsy==0.5.2
platformdirs==2.4.1
pycryptodome==3.12.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
requests-cache==0.5.2
scipy==1.7.3
six==1.16.0
statsmodels==0.13.1
stocktrends==0.1.5
tenacity==8.0.1
tomli==1.2.3
trading-ig==0.0.15
typing-extensions==4.0.1
urllib3==1.26.7
